feat: Organização do código responsável por se conectar ao Wi-Fi e criar a página Web
