+++ CMakeLists.txt
@@ -27,6 +27,7 @@ include_directories(${CMAKE_SOURCE_DIR}/libs/include) # Adiciona o diretório 'l
 
 add_executable(${PROJECT_NAME}  
         main.c # Código principal em C
+        libs/src/conexao.c 
         )
 
 target_include_directories(${PROJECT_NAME} PRIVATE ${CMAKE_SOURCE_DIR})  # Adiciona o diretório raiz como include privado apenas para o target atual
